ml_backend/requirements.txt
fastapi
uvicorn
numpy
scikit-learn
joblib
google-generativeai
python-dotenv
httpx
